feat#18 : ServiceModule 분리
